refactor: refactor proxy module
