+++ server/src/proxy/grpc/mod.rs
@@ -1,5 +1,5 @@
 // Copyright 2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
 mod route;
-pub(crate) mod sql_query;
+mod sql_query;
 mod write;
