+++ server/src/proxy/grpc/sql_query.rs
@@ -4,12 +4,18 @@
 
 use std::time::Instant;
 
-use ceresdbproto::storage::{
-    storage_service_client::StorageServiceClient, SqlQueryRequest, SqlQueryResponse,
+use arrow_ext::ipc::{CompressOptions, CompressionMethod, RecordBatchesEncoder};
+use ceresdbproto::{
+    common::ResponseHeader,
+    storage::{
+        arrow_payload, sql_query_response, storage_service_client::StorageServiceClient,
+        ArrowPayload, SqlQueryRequest, SqlQueryResponse,
+    },
 };
-use common_types::request_id::RequestId;
+use common_types::{record_batch::RecordBatch, request_id::RequestId};
 use common_util::{error::BoxError, time::InstantExt};
 use futures::FutureExt;
+use http::StatusCode;
 use interpreters::{context::Context as InterpreterContext, factory::Factory, interpreter::Output};
 use log::{error, info, warn};
 use query_engine::executor::Executor as QueryExecutor;
@@ -22,25 +28,38 @@ use sql::{
 use tonic::{transport::Channel, IntoRequest};
 
 use crate::proxy::{
-    error::{BadRequest, BadRequestWithoutErr, Internal, InternalWithoutErr, Result},
+    error::{self, ErrNoCause, ErrWithCause, Result},
     forward::{ForwardRequest, ForwardResult},
     Context, Proxy,
 };
 
-pub enum QueryResponse {
-    Forward(SqlQueryResponse),
-    Direct(Output),
-}
-
 impl<Q: QueryExecutor + 'static> Proxy<Q> {
-    pub async fn handle_query(&self, ctx: Context, req: SqlQueryRequest) -> Result<QueryResponse> {
+    pub async fn handle_sql_query(&self, ctx: Context, req: SqlQueryRequest) -> SqlQueryResponse {
+        let ret = self.handle_query_internal(ctx, req).await;
+        let mut resp = SqlQueryResponse::default();
+        match ret {
+            Err(e) => {
+                resp.header = Some(error::build_err_header(e));
+            }
+            Ok(v) => {
+                resp = v;
+            }
+        }
+        resp
+    }
+
+    async fn handle_query_internal(
+        &self,
+        ctx: Context,
+        req: SqlQueryRequest,
+    ) -> Result<SqlQueryResponse> {
         let req = match self.maybe_forward_query(&req).await {
-            Some(resp) => return Ok(QueryResponse::Forward(resp?)),
+            Some(resp) => return resp,
             None => req,
         };
 
         let output = self.fetch_query_output(ctx, &req).await?;
-        Ok(QueryResponse::Direct(output))
+        convert_output(&output, self.resp_compress_min_length)
     }
 
     async fn maybe_forward_query(&self, req: &SqlQueryRequest) -> Option<Result<SqlQueryResponse>> {
@@ -68,7 +87,8 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
                     .await
                     .map(|resp| resp.into_inner())
                     .box_err()
-                    .context(Internal {
+                    .context(ErrWithCause {
+                        code: StatusCode::INTERNAL_SERVER_ERROR,
                         msg: "Forwarded query failed",
                     })
             }
@@ -119,13 +139,15 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
         let mut stmts = frontend
             .parse_sql(&mut sql_ctx, &req.sql)
             .box_err()
-            .context(BadRequest {
+            .context(ErrWithCause {
+                code: StatusCode::BAD_REQUEST,
                 msg: "failed to parse sql",
             })?;
 
         ensure!(
             !stmts.is_empty(),
-            BadRequestWithoutErr {
+            ErrNoCause {
+                code: StatusCode::BAD_REQUEST,
                 msg: format!("No valid query statement provided, sql:{}", req.sql),
             }
         );
@@ -134,7 +156,8 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
         // TODO(yingwen): INSERT/UPDATE/DELETE can be batched
         ensure!(
             stmts.len() == 1,
-            BadRequestWithoutErr {
+            ErrNoCause {
+                code: StatusCode::BAD_REQUEST,
                 msg: format!(
                     "Only support execute one statement now, current num:{}, sql:{}",
                     stmts.len(),
@@ -150,7 +173,8 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
             // return internal server error in those cases
             .statement_to_plan(&mut sql_ctx, stmts.remove(0))
             .box_err()
-            .with_context(|| Internal {
+            .with_context(|| ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
                 msg: format!("Failed to create plan, query:{}", req.sql),
             })?;
 
@@ -158,13 +182,15 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
             .limiter
             .try_limit(&plan)
             .box_err()
-            .context(Internal {
+            .context(ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
                 msg: "Query is blocked",
             })?;
 
         if let Some(deadline) = deadline {
             if deadline.check_deadline() {
-                return InternalWithoutErr {
+                return ErrNoCause {
+                    code: StatusCode::INTERNAL_SERVER_ERROR,
                     msg: "Query timeout",
                 }
                 .fail();
@@ -185,7 +211,8 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
         let interpreter = interpreter_factory
             .create(interpreter_ctx, plan)
             .box_err()
-            .with_context(|| Internal {
+            .with_context(|| ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
                 msg: "Failed to create interpreter",
             })?;
 
@@ -196,14 +223,16 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
             )
             .await
             .box_err()
-            .context(Internal {
+            .context(ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
                 msg: "Query timeout",
             })?
         } else {
             interpreter.execute().await
         }
         .box_err()
-        .with_context(|| Internal {
+        .with_context(|| ErrWithCause {
+            code: StatusCode::INTERNAL_SERVER_ERROR,
             msg: format!("Failed to execute interpreter, sql:{}", req.sql),
         })?;
 
@@ -217,5 +246,127 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
     );
 
         Ok(output)
+    }
+}
+
+// TODO(chenxiang): Output can have both `rows` and `affected_rows`
+pub fn convert_output(
+    output: &Output,
+    resp_compress_min_length: usize,
+) -> Result<SqlQueryResponse> {
+    match output {
+        Output::Records(batches) => {
+            let mut writer = QueryResponseWriter::new(resp_compress_min_length);
+            writer.write_batches(batches)?;
+            writer.finish()
+        }
+        Output::AffectedRows(rows) => {
+            Ok(QueryResponseBuilder::with_ok_header().build_with_affected_rows(*rows))
+        }
+    }
+}
+
+/// Builder for building [`SqlQueryResponse`].
+#[derive(Debug, Default)]
+pub struct QueryResponseBuilder {
+    header: ResponseHeader,
+}
+
+impl QueryResponseBuilder {
+    pub fn with_ok_header() -> Self {
+        let header = ResponseHeader {
+            code: StatusCode::OK.as_u16() as u32,
+            ..Default::default()
+        };
+        Self { header }
+    }
+
+    pub fn build_with_affected_rows(self, affected_rows: usize) -> SqlQueryResponse {
+        let output = Some(sql_query_response::Output::AffectedRows(
+            affected_rows as u32,
+        ));
+        SqlQueryResponse {
+            header: Some(self.header),
+            output,
+        }
+    }
+
+    pub fn build_with_empty_arrow_payload(self) -> SqlQueryResponse {
+        let payload = ArrowPayload {
+            record_batches: Vec::new(),
+            compression: arrow_payload::Compression::None as i32,
+        };
+        self.build_with_arrow_payload(payload)
+    }
+
+    pub fn build_with_arrow_payload(self, payload: ArrowPayload) -> SqlQueryResponse {
+        let output = Some(sql_query_response::Output::Arrow(payload));
+        SqlQueryResponse {
+            header: Some(self.header),
+            output,
+        }
+    }
+}
+
+/// Writer for encoding multiple [`RecordBatch`]es to the [`SqlQueryResponse`].
+///
+/// Whether to do compression depends on the size of the encoded bytes.
+pub struct QueryResponseWriter {
+    encoder: RecordBatchesEncoder,
+}
+
+impl QueryResponseWriter {
+    pub fn new(compress_min_length: usize) -> Self {
+        let compress_opts = CompressOptions {
+            compress_min_length,
+            method: CompressionMethod::Zstd,
+        };
+        Self {
+            encoder: RecordBatchesEncoder::new(compress_opts),
+        }
+    }
+
+    pub fn write(&mut self, batch: &RecordBatch) -> Result<()> {
+        self.encoder
+            .write(batch.as_arrow_record_batch())
+            .box_err()
+            .with_context(|| ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
+                msg: "failed to encode record batch".to_string(),
+            })
+    }
+
+    pub fn write_batches(&mut self, record_batch: &[RecordBatch]) -> Result<()> {
+        for batch in record_batch {
+            self.write(batch)?;
+        }
+
+        Ok(())
+    }
+
+    pub fn finish(self) -> Result<SqlQueryResponse> {
+        let compress_output = self
+            .encoder
+            .finish()
+            .box_err()
+            .with_context(|| ErrWithCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
+                msg: "failed to encode record batch".to_string(),
+            })?;
+
+        if compress_output.payload.is_empty() {
+            return Ok(QueryResponseBuilder::with_ok_header().build_with_empty_arrow_payload());
+        }
+
+        let compression = match compress_output.method {
+            CompressionMethod::None => arrow_payload::Compression::None,
+            CompressionMethod::Zstd => arrow_payload::Compression::Zstd,
+        };
+        let resp = QueryResponseBuilder::with_ok_header().build_with_arrow_payload(ArrowPayload {
+            record_batches: vec![compress_output.payload],
+            compression: compression as i32,
+        });
+
+        Ok(resp)
     }
 }
